pyproject.toml
[tool.poetry]
name = "llmdm"
version = "0.1.0"
description = ""
authors = ["Evin O'Shea <evinoshea@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
llmdm = "llmdm.game:run"
llmdm-debug = "llmdm.game:run_debug"

[tool.poetry.dependencies]
python = "^3.10"
accelerate = "*"
python-arango = "*"
transformers = "*"
torch = {version = "*", source = "pytorch"}
aqlm = {extras = ["gpu"], version = "^1.1.6"}
mock = "^5.1.0"
pytest = "^8.3.3"
timeout-decorator = "^0.5.0"
rapidfuzz = "^3.10.0"
opensearch-py = "^2.7.1"
protobuf = "^5.28.2"
tiktoken = "^0.8.0"
blobfile = "^3.0.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
